+++ src/query/service/src/pipelines/builders/builder_aggregate.rs
@@ -21,6 +21,7 @@ use common_expression::with_mappedhash_method;
 use common_expression::AggregateFunctionRef;
 use common_expression::DataBlock;
 use common_expression::DataSchemaRef;
+use common_expression::HashMethodKind;
 use common_functions::aggregates::AggregateFunctionFactory;
 use common_pipeline_core::processors::processor::ProcessorPtr;
 use common_pipeline_core::query_spill_prefix;
@@ -31,11 +32,18 @@ use common_sql::executor::AggregateExpand;
 use common_sql::executor::AggregateFinal;
 use common_sql::executor::AggregateFunctionDesc;
 use common_sql::executor::AggregatePartial;
+use common_sql::executor::PhysicalPlan;
 use common_sql::IndexType;
 use common_storage::DataOperator;
 
+use crate::pipelines::processors::transforms::build_partition_bucket;
+use crate::pipelines::processors::transforms::AggregateInjector;
 use crate::pipelines::processors::transforms::FinalSingleStateAggregator;
 use crate::pipelines::processors::transforms::PartialSingleStateAggregator;
+use crate::pipelines::processors::transforms::TransformAggregateSpillWriter;
+use crate::pipelines::processors::transforms::TransformGroupBySpillWriter;
+use crate::pipelines::processors::transforms::TransformPartialAggregate;
+use crate::pipelines::processors::transforms::TransformPartialGroupBy;
 use crate::pipelines::processors::AggregatorParams;
 use crate::pipelines::processors::TransformExpandGroupingSets;
 use crate::pipelines::PipelineBuilder;
