+++ src/query/service/src/pipelines/builders/builder_filter.rs
@@ -1,0 +1,74 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use common_exception::ErrorCode;
+use common_exception::Result;
+use common_expression::type_check::check_function;
+use common_functions::BUILTIN_FUNCTIONS;
+use common_pipeline_core::processors::processor::ProcessorPtr;
+use common_pipeline_transforms::processors::profile_wrapper::TransformProfileWrapper;
+use common_pipeline_transforms::processors::transforms::Transformer;
+use common_sql::evaluator::BlockOperator;
+use common_sql::evaluator::CompoundBlockOperator;
+use common_sql::executor::Filter;
+
+use crate::pipelines::PipelineBuilder;
+
+impl PipelineBuilder {
+    pub(crate) fn build_filter(&mut self, filter: &Filter) -> Result<()> {
+        self.build_pipeline(&filter.input)?;
+
+        let predicate = filter
+            .predicates
+            .iter()
+            .map(|expr| expr.as_expr(&BUILTIN_FUNCTIONS))
+            .try_reduce(|lhs, rhs| {
+                check_function(None, "and_filters", &[], &[lhs, rhs], &BUILTIN_FUNCTIONS)
+            })
+            .transpose()
+            .unwrap_or_else(|| {
+                Err(ErrorCode::Internal(
+                    "Invalid empty predicate list".to_string(),
+                ))
+            })?;
+
+        let num_input_columns = filter.input.output_schema()?.num_fields();
+        self.main_pipeline.add_transform(|input, output| {
+            let transform = CompoundBlockOperator::new(
+                vec![BlockOperator::Filter {
+                    projections: filter.projections.clone(),
+                    expr: predicate.clone(),
+                }],
+                self.func_ctx.clone(),
+                num_input_columns,
+            );
+
+            if self.enable_profiling {
+                Ok(ProcessorPtr::create(TransformProfileWrapper::create(
+                    transform,
+                    input,
+                    output,
+                    filter.plan_id,
+                    self.proc_profs.clone(),
+                )))
+            } else {
+                Ok(ProcessorPtr::create(Transformer::create(
+                    input, output, transform,
+                )))
+            }
+        })?;
+
+        Ok(())
+    }
+}
