+++ src/query/service/src/pipelines/builders/builder_union_all.rs
@@ -1,0 +1,96 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use async_channel::Receiver;
+use common_exception::Result;
+use common_expression::DataBlock;
+use common_pipeline_core::processors::processor::ProcessorPtr;
+use common_pipeline_sinks::UnionReceiveSink;
+use common_pipeline_transforms::processors::profile_wrapper::ProcessorProfileWrapper;
+use common_sql::executor::PhysicalPlan;
+use common_sql::executor::UnionAll;
+
+use crate::pipelines::processors::transforms::TransformMergeBlock;
+use crate::pipelines::PipelineBuilder;
+use crate::sessions::QueryContext;
+
+impl PipelineBuilder {
+    pub fn build_union_all(&mut self, union_all: &UnionAll) -> Result<()> {
+        self.build_pipeline(&union_all.left)?;
+        let union_all_receiver = self.expand_union_all(&union_all.right, union_all)?;
+        self.main_pipeline
+            .add_transform(|transform_input_port, transform_output_port| {
+                let transform = TransformMergeBlock::try_create(
+                    transform_input_port,
+                    transform_output_port,
+                    union_all.left.output_schema()?,
+                    union_all.right.output_schema()?,
+                    union_all.pairs.clone(),
+                    union_all_receiver.clone(),
+                )?;
+
+                if self.enable_profiling {
+                    Ok(ProcessorPtr::create(ProcessorProfileWrapper::create(
+                        transform,
+                        union_all.plan_id,
+                        self.proc_profs.clone(),
+                    )))
+                } else {
+                    Ok(ProcessorPtr::create(transform))
+                }
+            })?;
+        Ok(())
+    }
+
+    fn expand_union_all(
+        &mut self,
+        input: &PhysicalPlan,
+        union_plan: &UnionAll,
+    ) -> Result<Receiver<DataBlock>> {
+        let union_ctx = QueryContext::create_from(self.ctx.clone());
+        let mut pipeline_builder = PipelineBuilder::create(
+            self.func_ctx.clone(),
+            self.settings.clone(),
+            union_ctx,
+            self.enable_profiling,
+            self.proc_profs.clone(),
+        );
+        pipeline_builder.cte_state = self.cte_state.clone();
+        let mut build_res = pipeline_builder.finalize(input)?;
+
+        assert!(build_res.main_pipeline.is_pulling_pipeline()?);
+
+        let (tx, rx) = async_channel::unbounded();
+
+        build_res.main_pipeline.add_sink(|input_port| {
+            let transform =
+                UnionReceiveSink::create(Some(tx.clone()), input_port, self.ctx.clone());
+
+            if self.enable_profiling {
+                Ok(ProcessorPtr::create(ProcessorProfileWrapper::create(
+                    transform,
+                    union_plan.plan_id,
+                    self.proc_profs.clone(),
+                )))
+            } else {
+                Ok(ProcessorPtr::create(transform))
+            }
+        })?;
+
+        self.pipelines.push(build_res.main_pipeline);
+        self.pipelines
+            .extend(build_res.sources_pipelines.into_iter());
+        Ok(rx)
+    }
+}
